画面パスとapiルートを定数化
