+++ src/app/calendar/menu/[date]/edit/page.tsx
@@ -3,17 +3,25 @@
 import { useEffect, useState } from "react";
 import { useParams } from "next/navigation";
 import MenuForm from "../../../../../../features/calendar/components/Common/MenuForm";
+import { Session } from "next-auth";
+import { API } from "@/constants/api";
 
-export default function MenuEditPage() {
+type Props = {
+  session: Session;
+};
+
+export default function MenuEditPage({ session }: Props) {
     const { date: rawDate } = useParams();
     const date = Array.isArray(rawDate) ? rawDate[0] : rawDate ?? "";
     const [initialData, setInitialData] = useState(null);
     const [error, setError] = useState<string | null>(null);
+    const userId = Number(session.user.id);
+
 
     useEffect(() => {
         const fetchMenu = async () => {
             try {
-                const res = await fetch(`/api/menu/by-date/${date}`);
+                const res = await fetch(API.MENU.GET_BY_USER_ID('by-date', userId, date));
                 if (!res.ok) {
                     const errorData = await res.json();
                         setError(errorData.error ?? "不明なエラーが発生しました");
