+++ src/app/calendar/page.tsx
@@ -2,6 +2,6 @@ import { requireLogin } from "@/lib/requireLogin";
 import CalendarClient from "../../../features/calendar/components/CalendarClient";
 
 export default async function CalendarPage() {
-  const session = await requireLogin(); // ✅ サーバーでセッションチェック
+  const session = await requireLogin(); // サーバーでセッションチェック
   return <CalendarClient session={session} />;
 }
