+++ src/app/components/auth/SubmitButton.tsx
@@ -1,4 +1,5 @@
 import { Heart } from "lucide-react";
+import { usePathname } from "next/navigation";
 
 type Props = {
     handleSubmit: () => void,
@@ -6,6 +7,11 @@ type Props = {
 }
 
 export default function SubmitButton({ handleSubmit, isLoading}: Props) {
+    const pathname = usePathname();
+    const commonLabel = 
+        pathname === '/register'
+            ? 'アカウントを作成'
+            : 'ログイン';
     return (
         <button
             onClick={handleSubmit}
@@ -17,8 +23,7 @@ export default function SubmitButton({ handleSubmit, isLoading}: Props) {
             ) : (
             <>
                 <Heart className="w-5 h-5 mr-2" />
-                ログイン
-                {/* アカウントを作成 */}
+                {commonLabel}
             </>
             )}
         </button>
