+++ src/app/components/auth/SwitchAuthLink.tsx
@@ -1,4 +1,5 @@
 // 新規登録へのリンク
+import { PATHS } from "@/constants/paths";
 import { usePathname } from "next/navigation";
 
 export default function SwitchAuthLink() {
@@ -8,13 +9,13 @@ export default function SwitchAuthLink() {
           ? 'すでにアカウントをお持ちの方は '
           : 'アカウントをお持ちでない方は ';
     const linkLabel = 
-        pathname === '/register'
+        pathname === PATHS.REGISTER
           ? 'ログイン '
           : 'こちらから新規登録 ';
     const route = 
-        pathname === '/register'
-          ? '/login'
-          : '/register';
+        pathname === PATHS.REGISTER
+          ? PATHS.LOGIN
+          : PATHS.REGISTER;
     
     return (
         <p className="text-center text-gray-600 mt-8">
