+++ src/lib/requireLogin.ts
@@ -1,4 +1,5 @@
 // src/app/lib/requireLogin.ts
+import { PATHS } from "@/constants/paths";
 import { authOptions } from "../../lib/authOptions"; // 認証オプション
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
@@ -6,7 +7,7 @@ import { redirect } from "next/navigation";
 export async function requireLogin() {
   const session = await getServerSession(authOptions);
   if (!session) {
-    redirect("/login");
+    redirect(PATHS.LOGIN);
   }
   return session;
 }
